Safari Support, SRC Change Detection, & Additional Resize Detection
Changed the way that the parent window registers itself with the child iframe so that it worked in Safari.  The new method seemed to work well in all browsers except maybe IE11 and older.  The change also helps in detecting if the source of the iFrame changes due to the user clicking a link or such inside of it, which could cause a content size change and causes the iFrame events to get lost.  Also added some additional (probably crummy) resize detection because I came across a use instance where some pages were still loading additional content via Ajax and the first height detection was not accurate.